+++ bin/stream_transactions.py
@@ -6,7 +6,7 @@ from async_v20 import OandaClient
 async def stream():
     async with OandaClient() as client:
         async for rsp in await client.stream_transactions():
-            print(rsp.dict())
+            print(rsp.dict(json=True))
 
 
 loop = asyncio.get_event_loop()